+++ game/systems/s_scene_transition.h
@@ -1,0 +1,12 @@
+//project verse, 2017-2021
+//by jose pazos perez
+//all rights reserved uwu
+
+#pragma once
+
+#include "config.h"
+
+namespace Verse::System::SceneTransition
+{
+    void handle(Config &c, Component::SceneTransition* transition);
+}