+++ main.cpp
@@ -9,6 +9,8 @@
 #include "game.h"
 #include "serialization.h"
 
+#include "r_textures.h"
+
 using namespace Verse;
 
 int main(int argc, const char * argv[]) {
@@ -50,6 +52,20 @@ int main(int argc, const char * argv[]) {
     Serialization::loadScene("test_scene_2", scene2, config);
     Serialization::loadPlayer(scene2, config);
     config.available_scenes.push_back(&scene2);
+    
+    EntityID test_transition = scene.createEntity("scene_transition");
+    Component::SceneTransition* t = scene.addComponent<Component::SceneTransition>(test_transition);
+    t->to_pos = Vec2(32, 64);
+    t->to_scene = &scene2;
+    Component::Collider* c = scene.addComponent<Component::Collider>(test_transition);
+    c->transform = Rect2(150, 120, 30, 30);
+    c->layer = Component::ColliderLayers::EVENT;
+    Component::Texture* tex = scene.addComponent<Component::Texture>(test_transition);
+    Graphics::Texture::loadTexture("res/graphics/palette_multi.png", tex);
+    tex->transform = Rect2(150, 120, 30, 30);
+    tex->offset.push_back(Vec2(0,0));
+    tex->layer.push_back(0);
+    
     
     config.active_camera = scene.getComponent<Component::Camera>(player);
     if (config.active_camera == nullptr)